#1898 rename lfric-specific psyad source files
