+++ src/psyclone/psyad/domain/lfric/__init__.py
@@ -37,7 +37,6 @@
 LFRic-specific support for PSyAD.
 '''
 
-from psyclone.psyad.domain.lfric.generate_lfric_adjoint import (
-    generate_lfric_adjoint)
-from psyclone.psyad.domain.lfric.generate_lfric_adjoint_harness import (
+from psyclone.psyad.domain.lfric.fric_adjoint import generate_lfric_adjoint
+from psyclone.psyad.domain.lfric.lfric_adjoint_harness import (
     generate_lfric_adjoint_harness)
