+++ src/psyclone/psyad/domain/lfric/lfric_adjoint.py
