+++ src/psyclone/tests/psyad/domain/lfric/test_lfric_adjoint_harness.py
